refactor(app): move i18n folder inside lib folder
