+++ packages/app/src/main.tsx
@@ -5,7 +5,7 @@ import { Cache } from 'three'
 
 import App from './App.tsx'
 // setup i18n
-import './i18n/config.ts'
+import './lib/i18n/config.ts'
 import './styles/global.css'
 
 // enable immer map/set support
