+++ packages/app/src/stores/editorStore.ts
@@ -2,7 +2,7 @@ import { merge } from 'lodash-es'
 import { create } from 'zustand'
 import { immer } from 'zustand/middleware/immer'
 
-import i18n from '@/i18n/config'
+import i18n from '@/lib/i18n/config'
 import { getLogger } from '@/services/loggerService'
 import { type Settings, getStoredSettings } from '@/services/settings'
 import type { DeepPartial, Number3Tuple, PartialNumber3Tuple } from '@/types'
